## What was the end-user or developer problem that led to this PR?

We should make `bundler/setup` as fast as possible.

## What is your fix for the problem, implemented in this PR?

Avoid unnecessary work of figuring out which specs should be kept locked from the lockfile and which ones should be discarded for resolving, whenever possible (in frozen mode, or when no changes detected).

This speeds up `bundler/setup` time of `rails/rails` Gemfile on my machine by about 30%.

```
✗ hyperfine 'BUNDLER_VERSION=2.3.12 ruby -rbundler/setup -e1' 'BUNDLER_VERSION=2.4.0.dev ruby -rbundler/setup -e1'
Benchmark 1: BUNDLER_VERSION=2.3.12 ruby -rbundler/setup -e1
  Time (mean ± σ):     368.9 ms ±  10.7 ms    [User: 309.4 ms, System: 46.9 ms]
  Range (min … max):   360.4 ms … 395.4 ms    10 runs
 
Benchmark 2: BUNDLER_VERSION=2.4.0.dev ruby -rbundler/setup -e1
  Time (mean ± σ):     267.3 ms ±   3.2 ms    [User: 212.3 ms, System: 44.8 ms]
  Range (min … max):   263.9 ms … 272.1 ms    10 runs
 
Summary
  'BUNDLER_VERSION=2.4.0.dev ruby -rbundler/setup -e1' ran
    1.38 ± 0.04 times faster than 'BUNDLER_VERSION=2.3.12 ruby -rbundler/setup -e1'
```

## Make sure the following tasks are checked

- [x] Describe the problem / feature
- [ ] Write [tests](https://github.com/rubygems/rubygems/blob/master/bundler/doc/development/PULL_REQUESTS.md#tests) for features and bug fixes
- [x] Write code to solve the problem
- [x] Make sure you follow the [current code style](https://github.com/rubygems/rubygems/blob/master/bundler/doc/development/PULL_REQUESTS.md#code-formatting) and [write meaningful commit messages without tags](https://github.com/rubygems/rubygems/blob/master/bundler/doc/development/PULL_REQUESTS.md#commit-messages)