@deivid-rodriguez 2022-04-28
There are big speed ups on `rails/rails` repository Gemfile because it's big, I suspect it should also be the case for @ngan's. But this is also noticiable (5-10% speed up) on a default Rails generated application 🎉:

```
hyperfine 'BUNDLER_VERSION=2.3.12 ruby -rbundler/setup -e1' 'BUNDLER_VERSION=2.4.0.dev ruby -rbundler/setup -e1'
Benchmark 1: BUNDLER_VERSION=2.3.12 ruby -rbundler/setup -e1
  Time (mean ± σ):     156.8 ms ±   1.3 ms    [User: 114.2 ms, System: 34.0 ms]
  Range (min … max):   155.2 ms … 161.1 ms    18 runs
 
Benchmark 2: BUNDLER_VERSION=2.4.0.dev ruby -rbundler/setup -e1
  Time (mean ± σ):     147.5 ms ±   0.7 ms    [User: 105.7 ms, System: 33.2 ms]
  Range (min … max):   146.5 ms … 149.4 ms    19 runs
 
Summary
  'BUNDLER_VERSION=2.4.0.dev ruby -rbundler/setup -e1' ran
    1.06 ± 0.01 times faster than 'BUNDLER_VERSION=2.3.12 ruby -rbundler/setup -e1'
```