@ngan 2022-04-28
Nice work, @deivid-rodriguez!

I ran some benchmarks for our gemfile.

`master`:
```
$ time bin/rubocop --version
bin/rubocop --version  2.63s user 0.40s system 97% cpu 3.115 total
```

`speed-up-bundler-setup`:
```
$ time bin/rubocop --version
bin/rubocop --version  1.96s user 0.37s system 93% cpu 2.486 total
```

So 25% improvement! It's good...but it's unfortunately still [not as good as subbundle](https://github.com/bigrails/subbundle#anecdotal-benchmarks).

With subbundle:
```
$ time bin/rubocop --version
bin/rubocop --version  0.51s user 0.16s system 92% cpu 0.720 total
```