@deivid-rodriguez 2022-04-29
Thanks for testing this out @ngan!

Your results are along the lines of what I expected. I think this effectively closes https://github.com/rubygems/rubygems/issues/5345, right?

Regarding the rest of the speed up, I think it probably can be obtained through an `only` option now? I recall the implementation of an `only` option was very simple, I think someone proposed it at some point. It would be good to get that rebased and ready, and perform some experiments with it to see how much it buys us.

In any case, I have some extra speed ups in mind that I can hopefully implement and will get to every user "for free".