@deivid-rodriguez 2022-04-29
@ngan Here's an extra patch that you can try out to see where it gets us:

```diff
diff --git a/bundler/lib/bundler/definition.rb b/bundler/lib/bundler/definition.rb
index 5c659608b0..ff4004fc11 100644
--- a/bundler/lib/bundler/definition.rb
+++ b/bundler/lib/bundler/definition.rb
@@ -146,7 +146,7 @@ def initialize(lockfile, dependencies, sources, unlock, ruby_version = nil, opti
       @dependency_changes = converge_dependencies
       @local_changes = converge_locals
 
-      @locked_specs_incomplete_for_platform = !@locked_specs.for(requested_dependencies & expand_dependencies(locked_dependencies), true, true)
+      @locked_specs_incomplete_for_platform = !@locked_specs.for(requested_dependencies & locked_dependencies, true, true)
 
       @requires = compute_requires
     end
@@ -245,10 +245,9 @@ def specs_for(groups)
 
     def dependencies_for(groups)
       groups.map!(&:to_sym)
-      deps = current_dependencies.reject do |d|
+      current_dependencies.reject do |d|
         (d.groups & groups).empty?
       end
-      expand_dependencies(deps)
     end
 
     # Resolve all the dependencies specified in Gemfile. It ensures that
```