@duckinator 2022-04-29
@deivid-rodriguez This looks really good, and is a surprisingly minor change for a noticeable speed improvement!

Are there any known/expected changes in behavior? Or things that need tests to be added? (None jumped out at me when I looked.)