@deivid-rodriguez 2022-05-01
@duckinator No, I don't think so, although I'm preparing other patches which in my testing improve performance of setting up `rails/rails` repository's Gemfile by about 100%. And those might actually bring a small change.

@ngan Yes, I'm expecting that too, given that you have a lot of dependencies, and `expand_dependencies` loops through all of them.