@ngan 2022-05-02
@deivid-rodriguez I applied [the second patch](https://github.com/rubygems/rubygems/pull/5503#issuecomment-1113106680) on top of the first. Unfortunately, it didn't yield any improvements.

```
$ time bin/rubocop --version
bin/rubocop --version  2.00s user 0.38s system 95% cpu 2.477 total
```