@deivid-rodriguez 2022-05-03
Too bad! Anyways, I'll merge the current speed up and iterate on this in future versions!